2014-12-03 11:46 UTC+0100 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * include/hbexprb.c     ! fixed GPF in compilation of this code:          DO hb_i18n_getText
  * include/achoice.ch
  * include/blob.ch
  * include/dbinfo.ch
  * include/hbapi.h
  * include/hbapicdp.h
  * include/hbatomic.h
  * include/hbclass.ch
  * include/hbinit.h
  * include/hbsetup.ch
  * include/hbsetup.h
  * include/inkey.ch
  * include/ord.ch
  * include/simpleio.ch
  * src/common/hbver.c
  * src/rtl/hbi18n2.prg
  * src/rtl/val.c
    * synced with Viktor's branch (casing, formatting, comment updating,
      clang support, use hb_FReadLen(), use hb_GetStdIn(), removed redundant
      casting)

  * include/fileio.ch
    + added HB_FLX_* constants

  * include/harbour.hbx
    * added __wapi_GetConsoleOutputCP()

  * include/hbapi.h
  * src/common/hbver.c
    + added hb_iswin*() functions borrowed from Viktor's branch

+++ src/rtl/val.c
@@ -63,7 +63,7 @@ HB_FUNC( VAL )
       HB_MAXINT lValue;
       double dValue;
 
-      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue , &iDec, &iWidth );
+      fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue, &iDec, &iWidth );
 
       if( ! fDbl )
          hb_retnintlen( lValue, iWidth );
@@ -89,7 +89,7 @@ HB_FUNC( HB_VAL )
       fDbl = hb_valStrnToNum( szText, iLen, &lValue, &dValue , &iDec, &iWidth );
 
       if( HB_ISNUM( 2 ) )
-         iLen = ( int ) hb_parni( 2 );
+         iLen = hb_parni( 2 );
 
       if( fDbl && iDec > 0 )
          iLen -= iDec + 1;
